+++ offchain/crates/validator-debt/src/lib.rs
@@ -7,6 +7,7 @@ pub mod jito;
 pub mod ledger;
 pub mod rewards;
 pub mod rpc;
+pub mod s3_fetcher;
 pub mod solana_debt_calculator;
 pub mod transaction;
 pub mod validator_debt;